linear search

+++ 0x1E-search_algorithms/0-linear.c
@@ -12,14 +12,14 @@
  */
 int linear_search(int *array, size_t size, int value)
 {
-	int i;
+	size_t i;
 
 	if (array == NULL)
 		return (-1);
 
 	for (i = 0; i < size; i++){
 		if (array[i] == value)
-			return i;
+			return (int)i;
 		printf("Value checked array[%ld] = [%d]\n", i, array[i]);
 	}
 	return (1);
